fix: Moved AbstractPublicTest to PublicTestsSuite

rh-pre-commit.version: 2.2.0
rh-pre-commit.check-secrets: ENABLED
